requirements.txt
asgiref==3.8.0
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
